refactor(validations): split medication schemas into focused modules behind a barrel
